Move amortized provider map to provider map
