+++ koku/api/report/ocp/provider_map.py
@@ -1,9 +1,12 @@
 #
-# Copyright 2021 Red Hat Inc.
+# Copyright 2022 Red Hat Inc.
 # SPDX-License-Identifier: Apache-2.0
 #
 """Provider Mapper for OCP Reports."""
+from functools import cached_property
+
 from django.contrib.postgres.aggregates import ArrayAgg
+from django.db.models import Case
 from django.db.models import CharField
 from django.db.models import DecimalField
 from django.db.models import F
@@ -11,11 +14,11 @@ from django.db.models import Max
 from django.db.models import Q
 from django.db.models import Sum
 from django.db.models import Value
+from django.db.models import When
 from django.db.models.functions import Coalesce
 
 from api.models import Provider
 from api.report.provider_map import ProviderMap
-from koku.database import KeyDecimalTransform
 from providers.provider_access import ProviderAccessor
 from reporting.models import OCPUsageLineItemDailySummary
 from reporting.provider.ocp.models import OCPCostSummaryByNodeP
@@ -29,6 +32,88 @@ from reporting.provider.ocp.models import OCPVolumeSummaryP
 
 class OCPProviderMap(ProviderMap):
     """OCP Provider Map."""
+
+    def __cost_model_cost(self, cost_model_rate_type=None):
+        """Return ORM term for cost model cost"""
+        if cost_model_rate_type:
+            return Sum(
+                Case(
+                    When(
+                        cost_model_rate_type=cost_model_rate_type,
+                        then=Coalesce(F("cost_model_cpu_cost"), Value(0, output_field=DecimalField()))
+                        + Coalesce(F("cost_model_memory_cost"), Value(0, output_field=DecimalField()))
+                        + Coalesce(F("cost_model_volume_cost"), Value(0, output_field=DecimalField())),
+                    ),
+                    default=Value(0, output_field=DecimalField()),
+                )
+                * Coalesce("exchange_rate", Value(1, output_field=DecimalField())),
+            )
+        else:
+            return Sum(
+                (
+                    Coalesce(F("cost_model_cpu_cost"), Value(0, output_field=DecimalField()))
+                    + Coalesce(F("cost_model_memory_cost"), Value(0, output_field=DecimalField()))
+                    + Coalesce(F("cost_model_volume_cost"), Value(0, output_field=DecimalField()))
+                )
+                * Coalesce("exchange_rate", Value(1, output_field=DecimalField())),
+            )
+
+    def __cost_model_cpu_cost(self, cost_model_rate_type=None):
+        """Return ORM term for cost model cost"""
+        if cost_model_rate_type:
+            return Sum(
+                Case(
+                    When(
+                        cost_model_rate_type=cost_model_rate_type,
+                        then=Coalesce(F("cost_model_cpu_cost"), Value(0, output_field=DecimalField())),
+                    ),
+                    default=Value(0, output_field=DecimalField()),
+                )
+                * Coalesce("exchange_rate", Value(1, output_field=DecimalField())),
+            )
+        else:
+            return Sum(
+                Coalesce(F("cost_model_cpu_cost"), Value(0, output_field=DecimalField()))
+                * Coalesce("exchange_rate", Value(1, output_field=DecimalField())),
+            )
+
+    def __cost_model_memory_cost(self, cost_model_rate_type=None):
+        """Return ORM term for cost model cost"""
+        if cost_model_rate_type:
+            return Sum(
+                Case(
+                    When(
+                        cost_model_rate_type=cost_model_rate_type,
+                        then=Coalesce(F("cost_model_memory_cost"), Value(0, output_field=DecimalField())),
+                    ),
+                    default=Value(0, output_field=DecimalField()),
+                )
+                * Coalesce("exchange_rate", Value(1, output_field=DecimalField())),
+            )
+        else:
+            return Sum(
+                Coalesce(F("cost_model_memory_cost"), Value(0, output_field=DecimalField()))
+                * Coalesce("exchange_rate", Value(1, output_field=DecimalField())),
+            )
+
+    def __cost_model_volume_cost(self, cost_model_rate_type=None):
+        """Return ORM term for cost model cost"""
+        if cost_model_rate_type:
+            return Sum(
+                Case(
+                    When(
+                        cost_model_rate_type=cost_model_rate_type,
+                        then=Coalesce(F("cost_model_volume_cost"), Value(0, output_field=DecimalField())),
+                    ),
+                    default=Value(0, output_field=DecimalField()),
+                )
+                * Coalesce("exchange_rate", Value(1, output_field=DecimalField())),
+            )
+        else:
+            return Sum(
+                Coalesce(F("cost_model_volume_cost"), Value(0, output_field=DecimalField()))
+                * Coalesce("exchange_rate", Value(1, output_field=DecimalField())),
+            )
 
     def __init__(self, provider, report_type):
         """Constructor."""
@@ -59,540 +144,36 @@ class OCPProviderMap(ProviderMap):
                         "tables": {"query": OCPUsageLineItemDailySummary},
                         "aggregates": {
                             "sup_raw": Sum(Value(0, output_field=DecimalField())),
-                            "sup_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_usage": self.cost_model_supplementary_cost,
                             "sup_markup": Sum(Value(0, output_field=DecimalField())),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_total": self.cost_model_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost,
+                            "infra_usage": self.cost_model_infrastructure_cost,
+                            "infra_markup": self.markup_cost,
+                            "infra_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost,
+                            "cost_usage": self.cost_model_cost,
+                            "cost_markup": self.markup_cost,
+                            "cost_total": self.cloud_infrastructure_cost + self.markup_cost + self.cost_model_cost,
                         },
                         "default_ordering": {"cost_total": "desc"},
                         "annotations": {
-                            "sup_raw": Value(0, output_field=DecimalField()),
-                            "sup_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_markup": Value(0, output_field=DecimalField()),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            # Cost =  Supplementary[field] + Infrastructure[filed]
-                            # Note: if a value was currently zero it was left out, unless both sup & infra are zero
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_raw": Sum(Value(0, output_field=DecimalField())),
+                            "sup_usage": self.cost_model_supplementary_cost,
+                            "sup_markup": Sum(Value(0, output_field=DecimalField())),
+                            "sup_total": self.cost_model_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost,
+                            "infra_usage": self.cost_model_infrastructure_cost,
+                            "infra_markup": self.markup_cost,
+                            "infra_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost,
+                            "cost_usage": self.cost_model_cost,
+                            "cost_markup": self.markup_cost,
+                            "cost_total": self.cloud_infrastructure_cost + self.markup_cost + self.cost_model_cost,
                             # the `currency_annotation` is inserted by the `annotations` property of the query-handler
                             "cost_units": Coalesce("currency_annotation", Value("USD", output_field=CharField())),
                             "clusters": ArrayAgg(Coalesce("cluster_alias", "cluster_id"), distinct=True),
@@ -602,64 +183,7 @@ class OCPProviderMap(ProviderMap):
                         },
                         "capacity_aggregate": {},
                         "delta_key": {
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            )
+                            "cost_total": self.cloud_infrastructure_cost + self.markup_cost + self.cost_model_cost,
                         },
                         "filter": [{}],
                         "cost_units_key": "raw_currency",
@@ -669,566 +193,40 @@ class OCPProviderMap(ProviderMap):
                         "tables": {"query": OCPUsageLineItemDailySummary},
                         "aggregates": {
                             "sup_raw": Sum(Value(0, output_field=DecimalField())),
-                            "sup_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_usage": self.cost_model_supplementary_cost,
                             "sup_markup": Sum(Value(0, output_field=DecimalField())),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(
-                                    F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField())
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(
-                                        F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                    + Coalesce(
-                                        F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            # Cost =  Supplementary[field] + Infrastructure[filed]
-                            # Note: if a value was currently zero it was left out, unless both sup & infra are zero
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(
-                                    F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField())
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(
-                                        F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                    + Coalesce(
-                                        F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_total": self.cost_model_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost_by_project,
+                            "infra_usage": self.cost_model_infrastructure_cost,
+                            "infra_markup": self.markup_cost_by_project,
+                            "infra_total": self.cloud_infrastructure_cost_by_project
+                            + self.markup_cost_by_project
+                            + self.cost_model_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost_by_project,
+                            "cost_usage": self.cost_model_cost,
+                            "cost_markup": self.markup_cost_by_project,
+                            "cost_total": self.cloud_infrastructure_cost_by_project
+                            + self.markup_cost_by_project
+                            + self.cost_model_cost,
                         },
                         "default_ordering": {"cost_total": "desc"},
                         "annotations": {
-                            "sup_raw": Value(0, output_field=DecimalField()),
-                            "sup_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_markup": Value(0, output_field=DecimalField()),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(
-                                    F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField())
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(
-                                        F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                    + Coalesce(
-                                        F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            # Cost =  Supplementary[field] + Infrastructure[filed]
-                            # Note: if a value was currently zero it was left out, unless both sup & infra are zero
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(
-                                    F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField())
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(
-                                        F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                    + Coalesce(
-                                        F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_raw": Sum(Value(0, output_field=DecimalField())),
+                            "sup_usage": self.cost_model_supplementary_cost,
+                            "sup_markup": Sum(Value(0, output_field=DecimalField())),
+                            "sup_total": self.cost_model_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost_by_project,
+                            "infra_usage": self.cost_model_infrastructure_cost,
+                            "infra_markup": self.markup_cost_by_project,
+                            "infra_total": self.cloud_infrastructure_cost_by_project
+                            + self.markup_cost_by_project
+                            + self.cost_model_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost_by_project,
+                            "cost_usage": self.cost_model_cost,
+                            "cost_markup": self.markup_cost_by_project,
+                            "cost_total": self.cloud_infrastructure_cost_by_project
+                            + self.markup_cost_by_project
+                            + self.cost_model_cost,
                             # the `currency_annotation` is inserted by the `annotations` property of the query-handler
                             "cost_units": Coalesce("currency_annotation", Value("USD", output_field=CharField())),
                             "clusters": ArrayAgg(Coalesce("cluster_alias", "cluster_id"), distinct=True),
@@ -1238,68 +236,9 @@ class OCPProviderMap(ProviderMap):
                         },
                         "capacity_aggregate": {},
                         "delta_key": {
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(
-                                        F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                    + Coalesce(
-                                        F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField())
-                                    )
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_project_monthly_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "cost_total": self.cloud_infrastructure_cost_by_project
+                            + self.markup_cost_by_project
+                            + self.cost_model_cost,
                         },
                         "filter": [{}],
                         "cost_units_key": "raw_currency",
@@ -1308,134 +247,19 @@ class OCPProviderMap(ProviderMap):
                     "cpu": {
                         "aggregates": {
                             "sup_raw": Sum(Value(0, output_field=DecimalField())),
-                            "sup_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_usage": self.cost_model_cpu_supplementary_cost,
                             "sup_markup": Sum(Value(0, output_field=DecimalField())),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_total": self.cost_model_cpu_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost,
+                            "infra_usage": self.cost_model_cpu_infrastructure_cost,
+                            "infra_markup": self.markup_cost,
+                            "infra_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_cpu_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost,
+                            "cost_usage": self.cost_model_cpu_cost,
+                            "cost_markup": self.markup_cost,
+                            "cost_total": self.cloud_infrastructure_cost + self.markup_cost + self.cost_model_cpu_cost,
                             "usage": Sum("pod_usage_cpu_core_hours"),
                             "request": Sum("pod_request_cpu_core_hours"),
                             "limit": Sum("pod_limit_cpu_core_hours"),
@@ -1443,135 +267,20 @@ class OCPProviderMap(ProviderMap):
                         "capacity_aggregate": {"capacity": Max("cluster_capacity_cpu_core_hours")},
                         "default_ordering": {"usage": "desc"},
                         "annotations": {
-                            "sup_raw": Value(0, output_field=DecimalField()),
-                            "sup_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_markup": Value(0, output_field=DecimalField()),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_raw": Sum(Value(0, output_field=DecimalField())),
+                            "sup_usage": self.cost_model_cpu_supplementary_cost,
+                            "sup_markup": Sum(Value(0, output_field=DecimalField())),
+                            "sup_total": self.cost_model_cpu_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost,
+                            "infra_usage": self.cost_model_cpu_infrastructure_cost,
+                            "infra_markup": self.markup_cost,
+                            "infra_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_cpu_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost,
+                            "cost_usage": self.cost_model_cpu_cost,
+                            "cost_markup": self.markup_cost,
+                            "cost_total": self.cloud_infrastructure_cost + self.markup_cost + self.cost_model_cpu_cost,
                             # the `currency_annotation` is inserted by the `annotations` property of the query-handler
                             "cost_units": Coalesce("currency_annotation", Value("USD", output_field=CharField())),
                             "usage_units": Value("Core-Hours", output_field=CharField()),
@@ -1587,32 +296,7 @@ class OCPProviderMap(ProviderMap):
                         "delta_key": {
                             "usage": Sum("pod_usage_cpu_core_hours"),
                             "request": Sum("pod_request_cpu_core_hours"),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("cpu", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "cost_total": self.cloud_infrastructure_cost + self.markup_cost + self.cost_model_cpu_cost,
                         },
                         "filter": [{"field": "data_source", "operation": "exact", "parameter": "Pod"}],
                         "conditionals": {
@@ -1638,132 +322,21 @@ class OCPProviderMap(ProviderMap):
                     "memory": {
                         "aggregates": {
                             "sup_raw": Sum(Value(0, output_field=DecimalField())),
-                            "sup_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_usage": self.cost_model_memory_supplementary_cost,
                             "sup_markup": Sum(Value(0, output_field=DecimalField())),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                + Coalesce(
-                                    KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_total": self.cost_model_memory_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost,
+                            "infra_usage": self.cost_model_memory_infrastructure_cost,
+                            "infra_markup": self.markup_cost,
+                            "infra_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_memory_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost,
+                            "cost_usage": self.cost_model_memory_cost,
+                            "cost_markup": self.markup_cost,
+                            "cost_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_memory_cost,
                             "usage": Sum("pod_usage_memory_gigabyte_hours"),
                             "request": Sum("pod_request_memory_gigabyte_hours"),
                             "limit": Sum("pod_limit_memory_gigabyte_hours"),
@@ -1771,135 +344,22 @@ class OCPProviderMap(ProviderMap):
                         "capacity_aggregate": {"capacity": Max("cluster_capacity_memory_gigabyte_hours")},
                         "default_ordering": {"usage": "desc"},
                         "annotations": {
-                            "sup_raw": Value(0, output_field=DecimalField()),
-                            "sup_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_markup": Value(0, output_field=DecimalField()),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_raw": Sum(Value(0, output_field=DecimalField())),
+                            "sup_usage": self.cost_model_memory_supplementary_cost,
+                            "sup_markup": Sum(Value(0, output_field=DecimalField())),
+                            "sup_total": self.cost_model_memory_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost,
+                            "infra_usage": self.cost_model_memory_infrastructure_cost,
+                            "infra_markup": self.markup_cost,
+                            "infra_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_memory_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost,
+                            "cost_usage": self.cost_model_memory_cost,
+                            "cost_markup": self.markup_cost,
+                            "cost_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_memory_cost,
                             # the `currency_annotation` is inserted by the `annotations` property of the query-handler
                             "cost_units": Coalesce("currency_annotation", Value("USD", output_field=CharField())),
                             "usage": Sum("pod_usage_memory_gigabyte_hours"),
@@ -1915,32 +375,9 @@ class OCPProviderMap(ProviderMap):
                         "delta_key": {
                             "usage": Sum("pod_usage_memory_gigabyte_hours"),
                             "request": Sum("pod_request_memory_gigabyte_hours"),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("memory", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "cost_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_memory_cost,
                         },
                         "filter": [{"field": "data_source", "operation": "exact", "parameter": "Pod"}],
                         "conditionals": {
@@ -1967,269 +404,43 @@ class OCPProviderMap(ProviderMap):
                         "tag_column": "volume_labels",
                         "aggregates": {
                             "sup_raw": Sum(Value(0, output_field=DecimalField())),
-                            "sup_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_usage": self.cost_model_volume_supplementary_cost,
                             "sup_markup": Sum(Value(0, output_field=DecimalField())),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_total": self.cost_model_volume_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost,
+                            "infra_usage": self.cost_model_volume_infrastructure_cost,
+                            "infra_markup": self.markup_cost,
+                            "infra_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_volume_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost,
+                            "cost_usage": self.cost_model_volume_cost,
+                            "cost_markup": self.markup_cost,
+                            "cost_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_volume_cost,
                             "usage": Sum("persistentvolumeclaim_usage_gigabyte_months"),
                             "request": Sum("volume_request_storage_gigabyte_months"),
                             "capacity": Sum("persistentvolumeclaim_capacity_gigabyte_months"),
                         },
                         "default_ordering": {"usage": "desc"},
                         "annotations": {
-                            "sup_raw": Value(0, output_field=DecimalField()),
-                            "sup_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "sup_markup": Value(0, output_field=DecimalField()),
-                            "sup_total": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_usage": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_distributed": Sum(
-                                Coalesce(
-                                    KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                    Value(0, output_field=DecimalField()),
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "infra_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_raw": Sum(
-                                Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_usage": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_distributed": Sum(
-                                (
-                                    Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_markup": Sum(
-                                Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "sup_raw": Sum(Value(0, output_field=DecimalField())),
+                            "sup_usage": self.cost_model_volume_supplementary_cost,
+                            "sup_markup": Sum(Value(0, output_field=DecimalField())),
+                            "sup_total": self.cost_model_volume_supplementary_cost,
+                            "infra_raw": self.cloud_infrastructure_cost,
+                            "infra_usage": self.cost_model_volume_infrastructure_cost,
+                            "infra_markup": self.markup_cost,
+                            "infra_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_volume_infrastructure_cost,
+                            "cost_raw": self.cloud_infrastructure_cost,
+                            "cost_usage": self.cost_model_volume_cost,
+                            "cost_markup": self.markup_cost,
+                            "cost_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_volume_cost,
                             "usage": Sum("persistentvolumeclaim_usage_gigabyte_months"),
                             "request": Sum("volume_request_storage_gigabyte_months"),
                             "capacity": Sum("persistentvolumeclaim_capacity_gigabyte_months"),
@@ -2244,32 +455,9 @@ class OCPProviderMap(ProviderMap):
                         "delta_key": {
                             "usage": Sum("persistentvolumeclaim_usage_gigabyte_months"),
                             "request": Sum("volume_request_storage_gigabyte_months"),
-                            "cost_total": Sum(
-                                (
-                                    Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
-                                    + Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
-                                )
-                                * Coalesce("infra_exchange_rate", Value(1.0, output_field=DecimalField()))
-                                + (
-                                    Coalesce(
-                                        KeyDecimalTransform("storage", "supplementary_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("storage", "infrastructure_usage_cost"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "infrastructure_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                    + Coalesce(
-                                        KeyDecimalTransform("pvc", "supplementary_monthly_cost_json"),
-                                        Value(0, output_field=DecimalField()),
-                                    )
-                                )
-                                * Coalesce("exchange_rate", Value(1, output_field=DecimalField()))
-                            ),
+                            "cost_total": self.cloud_infrastructure_cost
+                            + self.markup_cost
+                            + self.cost_model_volume_cost,
                         },
                         "filter": [{"field": "data_source", "operation": "exact", "parameter": "Storage"}],
                         "cost_units_key": "raw_currency",
@@ -2315,3 +503,95 @@ class OCPProviderMap(ProviderMap):
             },
         }
         super().__init__(provider, report_type)
+
+    @cached_property
+    def cost_model_supplementary_cost(self):
+        """Return supplementary cost model costs."""
+        return self.__cost_model_cost(cost_model_rate_type="Supplementary")
+
+    @cached_property
+    def cost_model_infrastructure_cost(self):
+        """Return infrastructure cost model costs."""
+        return self.__cost_model_cost(cost_model_rate_type="Infrastructure")
+
+    @cached_property
+    def cost_model_cost(self):
+        """Return all cost model costs."""
+        return self.__cost_model_cost()
+
+    @cached_property
+    def cost_model_cpu_supplementary_cost(self):
+        """Return supplementary cost model costs."""
+        return self.__cost_model_cpu_cost(cost_model_rate_type="Supplementary")
+
+    @cached_property
+    def cost_model_cpu_infrastructure_cost(self):
+        """Return infrastructure cost model costs."""
+        return self.__cost_model_cpu_cost(cost_model_rate_type="Infrastructure")
+
+    @cached_property
+    def cost_model_cpu_cost(self):
+        """Return all cost model costs."""
+        return self.__cost_model_cpu_cost()
+
+    @cached_property
+    def cost_model_memory_supplementary_cost(self):
+        """Return supplementary cost model costs."""
+        return self.__cost_model_memory_cost(cost_model_rate_type="Supplementary")
+
+    @cached_property
+    def cost_model_memory_infrastructure_cost(self):
+        """Return infrastructure cost model costs."""
+        return self.__cost_model_memory_cost(cost_model_rate_type="Infrastructure")
+
+    @cached_property
+    def cost_model_memory_cost(self):
+        """Return all cost model costs."""
+        return self.__cost_model_memory_cost()
+
+    @cached_property
+    def cost_model_volume_supplementary_cost(self):
+        """Return supplementary cost model costs."""
+        return self.__cost_model_volume_cost(cost_model_rate_type="Supplementary")
+
+    @cached_property
+    def cost_model_volume_infrastructure_cost(self):
+        """Return infrastructure cost model costs."""
+        return self.__cost_model_volume_cost(cost_model_rate_type="Infrastructure")
+
+    @cached_property
+    def cost_model_volume_cost(self):
+        """Return all cost model costs."""
+        return self.__cost_model_volume_cost()
+
+    @cached_property
+    def cloud_infrastructure_cost(self):
+        """Return ORM term for cloud infra costs."""
+        return Sum(
+            Coalesce(F("infrastructure_raw_cost"), Value(0, output_field=DecimalField()))
+            * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
+        )
+
+    @cached_property
+    def cloud_infrastructure_cost_by_project(self):
+        """Return ORM term for cloud infra costs by project."""
+        return Sum(
+            Coalesce(F("infrastructure_project_raw_cost"), Value(0, output_field=DecimalField()))
+            * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
+        )
+
+    @cached_property
+    def markup_cost(self):
+        """Return ORM term for cloud infra markup."""
+        return Sum(
+            Coalesce(F("infrastructure_markup_cost"), Value(0, output_field=DecimalField()))
+            * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
+        )
+
+    @cached_property
+    def markup_cost_by_project(self):
+        """Return ORM term for cloud infra markup by project."""
+        return Sum(
+            Coalesce(F("infrastructure_project_markup_cost"), Value(0, output_field=DecimalField()))
+            * Coalesce("infra_exchange_rate", Value(1, output_field=DecimalField()))
+        )
