+++ koku/forecast/forecast.py
@@ -41,12 +41,10 @@ from api.report.gcp.openshift.provider_map import OCPGCPProviderMap
 from api.report.gcp.provider_map import GCPProviderMap
 from api.report.oci.provider_map import OCIProviderMap
 from api.report.ocp.provider_map import OCPProviderMap
-from api.report.ocp.provider_map_amortized import OCPProviderMap as OCPProviderMapAmortized
 from api.utils import DateHelper
 from api.utils import get_cost_type
 from cost_models.models import CostModel
 from cost_models.models import CostModelMap
-from masu.processor import enable_ocp_amortized_monthly_cost
 from reporting.provider.aws.models import AWSOrganizationalUnit
 
 
@@ -594,12 +592,6 @@ class OCPForecast(Forecast):
 
     provider = Provider.PROVIDER_OCP
     provider_map_class = OCPProviderMap
-
-    def __init__(self, query_params):
-        """Initialize to override provider map class."""
-        if enable_ocp_amortized_monthly_cost(query_params.request.user.customer.schema_name):
-            self.provider_map_class = OCPProviderMapAmortized
-        super().__init__(query_params)
 
     @cached_property
     def source_to_currency_map(self):
